requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
astroid==3.2.4
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
dill==0.3.8
distro==1.9.0
exceptiongroup==1.2.1
filelock==3.15.4
flake8==7.1.1
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.4.0
google-auth==2.30.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
importlib_metadata==8.0.0
isort==5.13.2
jax==0.4.13
jaxlib==0.4.13
jsonpatch==1.33
jsonpointer==3.0.0
keras==2.10.0
Keras-Preprocessing==1.1.2
langchain==0.2.6
langchain-core==0.2.10
langchain-text-splitters==0.2.2
langsmith==0.1.82
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
mccabe==0.7.0
ml-dtypes==0.2.0
multidict==6.0.5
numpy==1.24.4
oauthlib==3.2.2
openai==1.37.1
opt-einsum==3.3.0
orjson==3.10.5
packaging==24.1
pillow==10.3.0
platformdirs==4.2.2
protobuf==3.19.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.1
pydantic==2.7.4
pydantic_core==2.18.4
pyflakes==3.2.0
PyMySQL==1.1.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.10.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
tenacity==8.4.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tokenizers==0.13.3
tomli==2.0.1
tomlkit==0.13.2
torch==1.12.0+cpu
torchaudio==0.12.0+cpu
torchvision==0.13.0+cpu
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2
